Changed coaddition method to use a common grid
Rather than use the wavelength array of the reference file as a reference grid, the code now uses a common grid with the same wavelength spacing which covers the full range of usable wavelengths. This should mean the coadded spectrum doesn't cut off abruptly when the wavelength array for the reference file ends.

I also re-enabled the diagnostic plots (note: there's still an issue with the diagnostic plots for the G160M data).