Trying to separate the reference coadd comparison from the main program
After generating the coadded spectrum, the program currently plots the result along with a reference coadd of the same object (PG1011-040) to make it easier to compare the results. This won't be useful when running the program on other targets, so the goal with this branch is to move that comparison plot to a separate program which uses the FITS files from the main code.